Support creating a HTTP client object.
The MarkLogic XCC protocol does not support triple objects returned
from sem:sparql(), and makes it harder to run the different query
types on the MarkLogic database.

This is the start of an abstraction layer on top of the REST API to
make running the different query types easier. It is designed
around the needs of this plugin, so may be harder to extend to
support the other REST APIs in another project.